src/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.14"
gradio = "^5.9.1"
python-dotenv = "^1.0.1"
openai = "^1.59.3"
langchain-openai = "^0.2.14"
langchain-core = "^0.3.29"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
